Restructed the application
Took the Django Two Scoops approach, merged in a little bit of
Django-Kevin, and decided to keep the templates within each app. That
way, the templates can be overriden by the system accordingly.

The files need to be updated to point to the correct location, and a few
more configuration things are needed. A deployment/setup script would be
nice.